requirements.txt
aiohttp==3.9.1
aiofiles==23.2.1
jinja2==3.1.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
